refactored getRate and getRates endpoints into unit testable functions
